+++ cellpy/utils/batch_tools/batch_core.py
@@ -149,7 +149,7 @@ class Data(collections.UserDict):
     loaded in memory. In future version, it could be that the Data object
     will return a link allowing querying instead to save memory usage...
 
-    Remark that some cellpy (cellreader.CellpyData) function might not work if
+    Remark that some cellpy (cellreader.CellpyCell) function might not work if
     you have the raw-data in memory, but not summary data (if the cellpy function
     requires summary data or other settings not set as default).
     """
@@ -249,7 +249,7 @@ class BaseExperiment(metaclass=abc.ABCMeta):
         self.summary_frames = None
         self.cell_data_frames = dict()
         self.memory_dumped = dict()
-        self.parent_level = "CellpyData"
+        self.parent_level = "CellpyCell"
         self.log_level = "CRITICAL"
         self._data = None
         self._store_data_object = True
@@ -307,7 +307,7 @@ class BaseExperiment(metaclass=abc.ABCMeta):
         if not os.path.isfile(cellpy_file_name):
             raise IOError
 
-        cellpy_object = cellreader.CellpyData(initialize=True)
+        cellpy_object = cellreader.CellpyCell(initialize=True)
         step_table = helper.look_up_and_get(
             cellpy_file_name, prms._cellpyfile_step, max_cycle=max_cycle
         )
@@ -323,7 +323,7 @@ class BaseExperiment(metaclass=abc.ABCMeta):
     def _load_cellpy_file(self, file_name):
         # TODO: modify this so that it can select parts of the data (max_cycle etc)
         selector = dict()
-        cellpy_data = cellreader.CellpyData()
+        cellpy_data = cellreader.CellpyCell()
         if self.max_cycle:
             cellpy_data.overwrite_able = False
             selector["max_cycle"] = self.max_cycle