+++ examples/simple_reader_cellpy.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 """simple script for reading .res-files from arbin
 
-This script uses the cellpy.cellreader.CellpyData object.
+This script uses the cellpy.cellreader.CellpyCell object.
 """
 
 import os
@@ -42,7 +42,7 @@ for f in rawfiles:
 
 print(80 * "-")
 
-d = cellreader.CellpyData().from_raw(rawfiles)
+d = cellreader.CellpyCell().from_raw(rawfiles)
 d.set_mass(mass)
 d.make_step_table()
 d.make_summary()