+++ tests/conftest.py
@@ -33,7 +33,7 @@ def cellpy_data_instance():
 
     log.setup_logging(default_level="DEBUG", testing=True)
 
-    return cellreader.CellpyData()
+    return cellreader.CellpyCell()
 
 
 @pytest.fixture
@@ -46,7 +46,7 @@ def dataset(cellpy_data_instance) -> Data:
         logging.info(
             f"pytest fixture could not find {fdv.cellpy_file_path} - making it from raw and saving"
         )
-        a = cellreader.CellpyData()
+        a = cellreader.CellpyCell()
         a.from_raw(fdv.res_file_path)
         a.set_mass(1.0)
         a.make_summary(find_ocv=False, find_ir=True, find_end_voltage=True)